requirements.txt
alembic==1.10.3
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.0
fonttools==4.39.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
pip==22.3.1
psycopg2-binary==2.9.6
pydantic==1.10.7
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.19.1
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
tenacity==8.2.2
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
wheel==0.38.4
